+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog of nens-auth-client
 - Added requests_session.OAuth2Session to get a requests Session with an access token.
   If the access token expires the token is refreshed.
 
+- Added requests_session.OAuth2CCSession for fetching an access token in the
+  Client Credentials Grant.
+
 - Fixed library with Django 4.*.
 
 - Fixed unittests with authlib 0.15.5.